### 2026-08-27 #804 @Pierre9344
Presubmission inquiry: tarpolyglot
Submitting Author Name: Pierre Solomon
Submitting Author Github Handle: @Pierre9344
Repository: https://github.com/Pierre9344/tarpolyglot
Submission type: Pre-submission

---

```
Package: tarpolyglot
Title: Run 'Python', 'Julia', 'Rust', and 'C++' Inside 'targets' Pipeline Steps
Version: 0.3.0
    person(
      given = "Pierre",
      family = "Solomon",
      email = "pierre.solomon@laposte.net",
      comment = c(ORCID = "0000-0001-7187-2664")
    )
  )
Description: Adds target constructors that make it easy to use 'Python', 'Julia',
    'Rust', and 'C++' inside a 'targets' pipeline using 'reticulate', 'JuliaCall',
    'rextendr', and 'Rcpp'. Provides tar_target_py(), tar_target_jl(),
    tar_target_rs(), and tar_target_cpp() (with matching _raw() variants), each
    mirroring targets::tar_target() and targets::tar_target_raw(). 'Python'
    and 'Julia' steps run a script via a live interpreter with optional R pre-
    and post-scripts; 'Rust' and 'C++' steps compile functions and call them
    from an R post-script. Results are then returned either as converted R
    objects or as files written to disk (format = "file"). Dynamic branching,
    environment/version selection, a 'crew' controller for isolation, and the
    full set of tar_target_raw() arguments are supported.
    https://pierre9344.github.io/tarpolyglot/
    targets,
    reticulate,
    JuliaCall,
    crew,
    rextendr,
    Rcpp,
    callr,
    cli
    tarchetypes,
    withr,
    RcppArmadillo
SystemRequirements: The external language toolchains are optional and each is
    only required by its matching targets constructor: a Python installation for
    tar_target_py(), a Julia installation for tar_target_jl(), a Rust
    toolchain with Cargo for tar_target_rs() (on Windows, the GNU toolchain, to
    match R's ABI), and a C++ compiler reachable by R (e.g. via Rtools on
    Windows) for tar_target_cpp(). A pipeline that uses a single language needs
    only that language's toolchain.
Config/roxygen2/version: 8.1.0
```

- **Explain how and why the package falls under these categories (briefly, 1-2 sentences). Please note any areas you are unsure of:**

`tarpolyglot` is an extension of the `targets` package that allow add new constructors (`tar_target_py()`, `tar_target_jl()`, `tar_target_rs()`, `tar_target_cpp()`, plus `_raw()` variants) that enable the use of `Python`, `Julia`, `Rust`, and `C++` to realize computation inside `targets` pipelines. Each constructor handles the setup a multilingual step would otherwise need to be written by the `targets` users: binding the interpreter or preparing the compiler build, selecting the environment and language version, moving data across the R boundary, and declaring the script and its upstream targets as tracked dependencies so `targets` re-runs the step when they change.

- **If submitting a statistical package, have you already [incorporated documentation of standards into your code via the srr package](https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-srr)?**

Not applicable. `tarpolyglot` is not a statistical package.srr

- **Who is the target audience and what are scientific applications of this package?**

The audience is researchers and research software engineers who already (or want to) use `targets` for reproducible analysis in R, but whose analysis is not entirely in R. This is common in practice: a bioinformatics or ecology pipeline whose data wrangling and modelling are in R, but who to use tools written in other language. For example, use of packages only available in Python or Julia. or whose performance-critical inner loop is written in Rust or C++.

Without `tarpolyglot`, this setup has to be written by hand, and it is easy to get wrong in ways that quietly break reproducibility: the user has to bind the interpreter or drive the compiler, move data across the R boundary, and make sure `targets` actually treats the foreign-language script as a dependency. When it does not, change detection breaks: you edit the Python script and the pipeline happily serves a stale result. `tarpolyglot` turns each of these steps into an ordinary target and removes that boilerplate. Concretely:

- A dedicated constructor for each supported language : `tar_target_py()`, `tar_target_jl()`, `tar_target_rs()`, and `tar_target_cpp()` (plus `_raw()` variants). This allow a multilingual step to be declared like any other `targets` step and forwards the full set of `targets::tar_target_raw()` arguments.

- Scripts kept up to date by `targets` : a foreign-language script, and the helper files it imports, can be declared as tracked dependencies, so editing them invalidates the step and the pipeline re-runs exactly what it should instead of serving a stale result.

- Clean per-step environments via `crew` : `polyglot_controller()` sets up a `crew` controller (`tasks_max = 1`) that gives each Python or Julia step a fresh interpreter process. This is the only way around `reticulate` and `JuliaCall` binding a single interpreter per R session, so different steps in one pipeline can each use their own clean environment.

The design goal is incremental adoption: you add one multilingual step to an existing `targets` pipeline, without migrating the project to a different build tool and without needing administrator rights to install one. That last point is concrete for my own work, which runs on a shared computing cluster with restricted user permissions.

- **Are there other R packages that accomplish the same thing? If so, how does yours differ or meet [our criteria for best-in-category](https://ropensci.github.io/dev_guide/policies.html#overlap)?**

The closest package in intent is [`rixpress`](https://github.com/b-rodrigues/rixpress), which also builds reproducible polyglot pipelines. It delegates execution, dependency resolution, and environment management to Nix, which buys strong isolation and reproducibility guarantees at the cost of adopting Nix. However Nix may not be installable on all system (e.g. Windows system need to pass by WSL, and users with limited rights that use public computational server may not be able to use Nix) which forbid the use of `rixpress`. `tarpolyglot` is deliberately narrower and less prescriptive: it stays an extension of `targets` and use leaves environment management to the language-native tools users already have (virtualenv, venv, uv, Poetry, conda, juliaup, Julia projects, rustup). It is aimed at users who want to add multilingual steps to an existing `targets` pipeline, including on machines where Nix is unavailable or cannot be installed by a non-privileged user. The README contains a fuller comparison.

The language bridges themselves (`reticulate`, `JuliaCall`, `rextendr`, `Rcpp`) are dependencies, not competitors: `tarpolyglot` does not reimplement any interop. You can of course call them from a plain `tar_target()`. What the constructors add is the pipeline-level layer that is tedious and easy to get wrong by hand:

- upstream targets are wired in declaratively with `inputs`, so `targets` detects them as dependencies;

- scripts and imported helper files can be tracked, so editing them invalidates the step;

- interpreter/environment and language-version selection are arguments, and an explicit choice wins over an ambient `RETICULATE_PYTHON` or `JULIA_PROJECT` inherited from an IDE or a `crew` worker;

- `polyglot_controller()` is an helper that gives each step a fresh interpreter process, which is the only way around `reticulate` and `JuliaCall` binding one interpreter per session;

- the `tarpolyglot_map()` and other included pattern helpers compile a Rust crate or C++ library once per targets steps with multiple branches. They allow to reuse it across the compiled library across all steps' branches instead of recompiling in every branch.

- **(If applicable) Does your package comply with our [guidance around *Ethics, Data Privacy and Human Subjects Research*](https://devguide.ropensci.org/policies.html#ethics-data-privacy-and-human-subjects-research)?**

Not applicable. `tarpolyglot` is developer tooling: it collects, transmits, and stores no data of its own, contacts no remote service, and has no connection to human subjects research. It only runs code the user has written, on the user's own machine, within the user's own pipeline.

- **Any other questions or issues we should be aware of?:**

**Optional system dependencies :**

The package sits on four external toolchains (Python, Julia, a Rust toolchain with Cargo, and a C++ compiler reachable by R). All are optional and each is needed only by its matching constructor; a pipeline using one language needs only that language's toolchain, and this is declared in `SystemRequirements`. `toolchain_check()` is provided as a diagnostic that reports what is reachable *from a fresh worker process*, which is the question that actually matters for a pipeline.

**Testing strategy, given the package dependencies :**

Tests that need a live interpreter or compiler are gated behind a `TARPOLYGLOT_INTEGRATION` environment variable, so `R CMD check` on CRAN and on a contributor's laptop stays fast and offline, in line with the CRAN policy that long-running tests be made optional. Those tests do run in CI on both Ubuntu and Windows, and include end-to-end pipelines built with `tar_make()` rather than only direct calls to the internals. Test coverage measured with the integration tests enabled is currently 90.77% (as reported by Codecov); measured without them it is lower, because much of the package is by nature code that only executes when a real toolchain is present.

**Known limitation (documented in vignettes) :**

- `reticulate` and `JuliaCall` can bind only one interpreter per R session. `polyglot_controller()` (a `crew` controller with `tasks_max = 1`) is the documented way to get per-step isolation.

- As `tarpolyglot` depends on other R packages (e.g. `reticulate` for Python, `JuliaCall` for Julia, `rextendr` for Rust, and `Rcpp` for C++) any limitations of these packages will also be inherited by tarpolyglot. For example, Julia's `Threads.@threads` does not work through `JuliaCall`; `Distributed` does.

  These limitations are covered in the vignettes.

**tarpolyglot is already published on CRAN :**

The version `0.2.0` was recently published on [CRAN](https://cran.r-project.org/web/packages/tarpolyglot/index.html). Since then, I improved the package. The modifications include mainly the new C++ support via dedicated Rcpp based controllers and a new `tarpolyglot::toolchain_check()` function to help users discover and check the toolchains available on their system environment. I therefore increased the `tarpolyglot` version to `0.3.0` in the `DESCRIPTION` file and modified the `NEWS.md` file, but I didn't submit it to CRAN, or added a new release on github, yet due to CRAN policies and the possible modifications I could be asked to do for this submission.
Both the `main` and `dev` branch of my github repository correspond to this lattest 0.3.0 version.

## Use of Generative AI

- [x] Generative AI tools were used to produce some of the material in this submission.

If so, please describe usage, and include links to any relevant aspects of your repository. See [our blog post](https://ropensci.org/blog/2026/02/26/ropensci-ai-policy/) for background. (Explicit advice is not yet included in our *Dev Guide*; we are hoping to update very soon, and ask your cooperation and transparency in the meantime.)

Claude Code (Anthropic's Claude Opus 4.8, later updated to Opus 5) was used as an assistant throughout development, I reviewing, testing, and directing all changes made by the AI. The package's design is entirelly based on my decisions: the constructor API and its argument set, the decision to mirror `targets` steps, the per-step isolation approach via `crew`, the compile-once branching strategy, and the overall scope of the package were all human decisions, taken for reasons specific to my own research computing environment.

Concretely, AI assistance was used for:
- **Code reviewing :** Revision and improvement of the `tarpolyglot` functions.
- **Documentation :** Drafting and revising the roxygen documentation, including the shared "Script options" / "Script arguments" sections in [`R/shared-params.R`](https://github.com/Pierre9344/tarpolyglot/blob/main/R/shared-params.R), and drafting the vignettes, in particular [`vignettes/scripts.Rmd`](https://github.com/Pierre9344/tarpolyglot/blob/main/vignettes/scripts.Rmd) and `vignettes/cpp.Rmd`.
- **Tests :** Drafting a substantial part of the test suite, including the unit tests for the internal helpers and the end-to-end pipeline tests in [`tests/testthat/test-integration-pipeline.R`](https://github.com/Pierre9344/tarpolyglot/blob/main/tests/testthat/test-integration-pipeline.R). Every assertion was checked against the package's actual behaviour rather than accepted on trust; several drafts were rejected or rewritten because they encoded a platform quirk or tested a mock rather than the package.
- **Continuous integration :** Drafting and debugging the GitHub Actions workflows in [`.github/workflows/`](https://github.com/Pierre9344/tarpolyglot/tree/main/.github/workflows), including the toolchain setup for the integration and coverage jobs.
- **Package logo :** The hex sticker in `man/figures/` (and the derived favicons in `pkgdown/favicon/`) is a SVG generated by Claude Code based on a prompt. It contains no third-party logo artwork: every element is either an original shape or type set in a font.
