### 2026-08-27 #804 @mpadge
Thanks @Pierre9344 - please edit your opening comment and replace the initial header matter in the exact format from [the template](https://github.com/ropensci/software-review/blob/main/.github/ISSUE_TEMPLATE/B-submit-a-presubmission-inquiry.md?plain=1). Entries need to look like this:
```

Submitting Author Name: Name
Submitting Author Github Handle: @github_handle
Other Package Authors Github handles: (comma separated, delete if none) @github_handle1, @github_handle2
Repository:  https://repourl
Submission type: Pre-submission
Language: en

---
```

As stated in the initial instructions, the ``-type HTML delimiters are needed for our bot to identify these variables. Once you've edited and updated, you may call `@ropensci-review-bot check package` yourself to trigger the checks. Thanks!
