### 2026-08-27 be52eb7
correction following ropensci pre-submission automatic report
