+++ reference/tar_target_rs.md
@@ -209,7 +209,7 @@ for worked examples of all three and guidance on choosing.
 ## Examples
 
 ``` r
-if (FALSE) { # \dontrun{
+# Building a target does not run it, so these examples need no Rust toolchain.
 # scripts/square.rs:
 #   #[extendr]
 #   fn square(x: f64) -> f64 { x * x }
@@ -223,12 +223,90 @@ list(
     post_script = "scripts/post.R"
   )
 )
+#> [[1]]
+#> <tar_stem> 
+#>   name: rs_square 
+#>   description:  
+#>   command:
+#>     tarpolyglot::run_rs_step(script = "scripts/square.rs", 
+#>         post_script = "scripts/post.R", inputs = list(x = value), 
+#>         output = "object", files = NULL, dependencies = NULL, features = NULL, 
+#>         profile = NULL, toolchain = NULL) 
+#>   format: rds 
+#>   repository: local 
+#>   iteration method: vector 
+#>   error mode: stop 
+#>   memory mode: auto 
+#>   storage mode: worker 
+#>   retrieval mode: auto 
+#>   deployment mode: worker 
+#>   priority: 0 
+#>   resources:
+#>     list() 
+#>   cue:
+#>     seed: TRUE
+#>     file: TRUE
+#>     iteration: TRUE
+#>     repository: TRUE
+#>     format: TRUE
+#>     depend: TRUE
+#>     command: TRUE
+#>     mode: thorough 
+#>   packages:
+#>     tarpolyglot
+#>     stats
+#>     graphics
+#>     grDevices
+#>     utils
+#>     datasets
+#>     methods
+#>     base 
+#>   library:
+#>     NULL
 
 # The three ways to supply a script (see the "Script options" section):
 # 1. Literal path: untracked, editing the file does NOT re-run the step.
 tarpolyglot::tar_target_rs(
   name = demo_literal, script = "rs/step.rs", post_script = "R/post.R"
 )
+#> <tar_stem> 
+#>   name: demo_literal 
+#>   description:  
+#>   command:
+#>     tarpolyglot::run_rs_step(script = "rs/step.rs", post_script = "R/post.R", 
+#>         inputs = list(), output = "object", files = NULL, dependencies = NULL, 
+#>         features = NULL, profile = NULL, toolchain = NULL) 
+#>   format: rds 
+#>   repository: local 
+#>   iteration method: vector 
+#>   error mode: stop 
+#>   memory mode: auto 
+#>   storage mode: worker 
+#>   retrieval mode: auto 
+#>   deployment mode: worker 
+#>   priority: 0 
+#>   resources:
+#>     list() 
+#>   cue:
+#>     seed: TRUE
+#>     file: TRUE
+#>     iteration: TRUE
+#>     repository: TRUE
+#>     format: TRUE
+#>     depend: TRUE
+#>     command: TRUE
+#>     mode: thorough 
+#>   packages:
+#>     tarpolyglot
+#>     stats
+#>     graphics
+#>     grDevices
+#>     utils
+#>     datasets
+#>     methods
+#>     base 
+#>   library:
+#>     NULL
 
 # 2. tar_target_path(): tracked, editing the file DOES re-run the step.
 list(
@@ -239,6 +317,82 @@ list(
     post_script = "R/post.R"
   )
 )
+#> [[1]]
+#> <tar_stem> 
+#>   name: step_rs 
+#>   description:  
+#>   command:
+#>     "rs/step.rs" 
+#>   format: file 
+#>   repository: local 
+#>   iteration method: vector 
+#>   error mode: stop 
+#>   memory mode: auto 
+#>   storage mode: worker 
[…]

[…]
