### 2026-08-27 #804 @adamhsparks
hi, @Pierre9344, this looks really neat and is something that I've considered with my team's work.

I'll confer with the other editors and get back to you.

In the meantime, I know it's not an R package, but @/b-rodrigues has been working on [tlang](https://tstats-project.org/). Could you add a short comparison with this as well that compliments your evaluation of {rixexpress}?
