+++ pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   python: python.html
   rust: rust.html
   scripts: scripts.html
-last_built: 2026-08-27T17:35Z
+last_built: 2026-08-28T03:07Z
 urls:
   reference: https://pierre9344.github.io/tarpolyglot/reference
   article: https://pierre9344.github.io/tarpolyglot/articles
