### 2026-08-28 #804 @Pierre9344
Hi, @adamhsparks,

I updated the README to add the T comparison. Please note that I never used T myself so my understanding of it is based on the content of the website you sent previously and the [blog posts](https://brodrigues.co/) that @b-rodrigues published about T earlier this year.
