### 2026-08-29 5be0116
C++ vignette correction
