### 2026-09-02 #804 @adamhsparks
@Pierre9344, thanks for your patience and updates to the README. We have determined that {tarpolyglot} is in scope and welcome a formal submission for review. Please close this issue and open a review request as the next steps.
