### 2026-09-03 5013104
readme update to add information about C++ support
