+++ pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   python: python.html
   rust: rust.html
   scripts: scripts.html
-last_built: 2026-08-28T03:07Z
+last_built: 2026-09-03T05:04Z
 urls:
   reference: https://pierre9344.github.io/tarpolyglot/reference
   article: https://pierre9344.github.io/tarpolyglot/articles
