### 2026-09-03 #804 @Pierre9344
Thank you. I'm closing this now. I precise that I just pushed some modifications for the readme to add missing information about a constructor. I also modified the cpp.Rmd vignette related to this constructor (no modification of R code).
